Amended to work with other OpenStack services
The Rackspace OpenCloud libraries are built from OpenStack and thus support connecting to any OpenStack service, not just Rackspace Cloud. These changes  just wrap the initiation and identity logic up and add a couple of new variables, to allow for this client to be used to connect to other OpenStack services without further modification.